gpio: return -ENOTSUPP if debounce cannot be set
It appears some drivers are using gpio_set_debounce()
opportunistically, i.e. without knowing whether it works or
not. (Example: input/keyboard/gpio_keys.c) to account for
this use case, return -ENOTSUPP and do not print any
warnings in this case.

Took a round over the other gpio_set_debounce() consumers
to make sure that none of them are relying on the returned
error code to be something specific.

Change-Id: I10650a2373d47c34ac4328167b128366c5088b04
Signed-off-by: Linus Walleij <linus.walleij@linaro.org>
Signed-off-by: anupritaisno1 <www.anuprita804@gmail.com>