Revert "Revert "Revert "camera_v2: isp: Replace monotonic timestamp with boot_time"""
This commit is wrong and shouldn't be picked in first place. The
issue is that we picked an incorrect commit in HAL, which only
converts the timestamp in HAL3. Thus camera recording is broken
if some apps need to use HAL1.

Since this commit is useless actually, drop the patch in HAL and
here both.

This reverts commit 1b386d82549863cf936628908963b01f98236d23.

Change-Id: I1844b332167dfcaf41697171a29059e8a833bc44
Signed-off-by: anupritaisno1 <www.anuprita804@gmail.com>